Time to make this a new stable version. The usability fixes, system tray, color settings, SMIL enhancements and improved TV source is enough for now. Playing two or more streams simultaneously with SMIL, has to wait for 0.9.1.
CCMAIL: kde-i18n-doc@kde.org

In about two weeks, I release a stable version. As such, the strings are
freezed. Compared to the previous version, there are some changes on the
general page of the config dialog.
Also, I added some help in the handbook. But since this is still unfinished,
I would consider this low priority for translations.
I'll try to make a -rc1 version this day and announce it on kde-apps.

TIA,

Koos


svn path=/trunk/extragear/multimedia/kmplayer/; revision=433362