Setting the rootLayout in the ViewArea, for smil, is not flexable enough for eg. a .m3u file with two smil links. Now SMIL::Smil registers itself when activated via the PlaylistNotify callback. Call it eventlistener, because its only used for dispatching events in the tree.
Use the id field for smil too, saving some string comparisons

Setting the bgcolor of the video widget did only setPaletteBackgroundColor,
but should also do XSetWindowBackground for the embedded window. Code already
there, but not called.

Make use of the eventlistener changes to allow RSS::Item to paint the 
description (copied form SMIL::TextMediaType). Though this is quite ugly and
since I should add a infopanel for the Real plugin compat, this could be
an excuse to implement it. Also, only use half the viewing area, so if there
is a video, it still shows (no idea if podcast can have it btw.)

svn path=/trunk/extragear/multimedia/kmplayer/; revision=435572