Create a real InfoWindow class and take QTextBrowser as base (instead of QTextEdit), does have troubles getting a real link (not '://') out of it
Get rid of moveRect optimalization, simply increase repaint rate to 1/50s

Repaint direct, and not via the 'repaint(rect, false)' route. Actually
remaining from trying to use XShm for the pixmap to paint in before being
copied to the window background. Unfortunately I couldn't get it working
w/o using hacks and then it turned out to be slower .... The attempt was to
create a new paintdevice for this so it wouldn't affect the painting calls
I do now on QPainter. Although QPainter takes a QPaintDevice as argument to
paint on, under the hood it asumes it's a QPixmap or QWidget (or one has
to reimplement really lots, using the cmd() interface). The alternative to
derive from QPixmap and hackerly swap it hd/rendhd handle, which did work
but turned out to be slower instead ...

svn path=/trunk/extragear/multimedia/kmplayer/; revision=451370