Restore opening the ao/vo driver once, because it didn't help clearing the widget between movies (as it's the initialization of the movie that seems to paint this garbage) and we benefit from added config keys for these drivers. Eg. the alsa driver has some useful setting. Tried loading all drivers and unloading them too, but that slows down too much ..
Heh heh, finally figured out why I always get B&W on the second movie with a
playlist. According to
http://xinehq.de/index.php/faq#CONTRASTBRIGHTNESSSATURATION
I need to play w/ the color settings a bit, and that was always auto done
for the first movie by the dcop link, but not for 1+ items. Code was there
but at the wrong side of the stream creation call :-)

Clean away the KXinePlayerPrivate class.


svn path=/trunk/extragear/multimedia/kmplayer/; revision=452125