Improved prefetching of media content
Instead of prefetching the first level children of all groups, do this
only for par and for seq prefetch only the first child. This also
happens now recursive, so if a child is a par, all its children are prefetched
too. First child of anchor and children of priority class are prefetches as
well.
When the seq moves to next child, prefetch it's next sibling.

svn path=/trunk/extragear/multimedia/kmplayer/; revision=823686