Actively update a timings element freeze state
Intermediate solution towards a better timing traversal in SMIL.
Remove the hacky and inefficient keepContent with a FreezeStateUpdater
that is executed everytime a group child begin or finish state changes.

Although untested, fill hold is implemented too.
The freezing between seq children, as in
<seq>
  <img region="region1" src="flower.png"/>
  <img region="region1" src="water.png" begin="2s" dur="3s"/>
</seq>
where the flower should be shown 2s, works now as well.

svn path=/trunk/extragear/multimedia/kmplayer/; revision=823687