Redo Smil av pausing prevention, now done in the process class
Code pulled over from the maemo port, which looks less hacky and
gains a paused state we may use in the UI. The main culprit though seems
to be that defering the MediaType, caused by the buffering state of the
backend process, first paused the document and then adjusted its state.
Since all media elements listen to document pausing events, a pause
command was issued.

svn path=/trunk/extragear/multimedia/kmplayer/; revision=1258260