Fixup kphononplayer compile, also use plain X11 reparenting
What failed sofar is setting the xcb input mask on child widgets. This
means that mouse moves are not detecteced and thus for fullscreen it's a
bit hard to get out.
Tried the xcb_query_tree method, just like before with the Xlib variant.
But firstly, I still didn't get the events and worse, it broke npplayer
for some reasons.

Anyhow, getting there ..