Fix the change window attribute value arg, restore sub window tracking
The 'static' modifier turned out to be evil (I'll admit to have googled
for this code quote). Maybe I should have just used '&mask' for this
single value.
Discarding replies might be nice if not wanting to handle any failure
case, but it would have pointed me earlier to the bad access reply.
Anyhow, works now and restore the sub window tracking for mouse
movements so the controlbar shows again when hovering at the bottom of
the full screen.