Hide fullscreen window before reparinting to dock's central widget
Sometimes, esp. with the kphononplayer, instead of showing in the
application window after restoring from fullscreen, the video window
was floating. I suspect that is because the unsetting of the fullscreen
window flag, the WM is busy decorating and finding a good spot for the
video window. If it was reparented in that time, adding the decorations
will reparent the widget to a decorated root window.
But that's just guessing :)
When I see this again, next step might be to wait for the unmap event,
wait a little, then unset fullscreen etc.
See whether this works reliable for a while.