Reparent fullscreen window after xcb tells us the window is unmapped
Yet another attempt to fight sometimes the formally fullscreen window is
floating instead of being the central widget.

So we call setVisible(false) and unset the fullscreen flag. New now is
that a timer is started which checks the mapped status and only when
unmapped, do the setCentralWidget call.
Seem to help, but might be the delay or mapped state check too.